+++ frontend/app/chat/[chatId]/components/ActionsBar/components/ChatInput/components/ChatBar/components/MentionInput/MentionInput.tsx
@@ -1,7 +1,5 @@
-/* eslint-disable max-lines */
 import Editor from "@draft-js-plugins/editor";
-import { EditorState, getDefaultKeyBinding } from "draft-js";
-import { ReactElement, useEffect } from "react";
+import { ReactElement } from "react";
 import { useTranslation } from "react-i18next";
 
 import "@draft-js-plugins/mention/lib/plugin.css";
@@ -10,9 +8,6 @@ import "draft-js/dist/Draft.css";
 import { AddNewBrainButton } from "./components";
 import { BrainSuggestion } from "./components/BrainSuggestion";
 import { useMentionInput } from "./hooks/useMentionInput";
-
-import { MentionTriggerType } from "@/app/chat/[chatId]/components/ActionsBar/types";
-import { Popover } from "@radix-ui/react-popover";
 
 type MentionInputProps = {
   onSubmit: () => void;
@@ -27,106 +22,22 @@ export const MentionInput = ({
   const {
     mentionInputRef,
     MentionSuggestions,
+    keyBindingFn,
     editorState,
     onOpenChange,
     onSearchChange,
     open,
     plugins,
-    setEditorState,
     suggestions,
     onAddMention,
-    mentionItems,
-    insertMention,
+    handleEditorChange,
   } = useMentionInput({
     message,
+    onSubmit,
+    setMessage,
   });
 
   const { t } = useTranslation(["chat"]);
-
-  type TriggerMap = {
-    trigger: MentionTriggerType;
-    content: string;
-  };
-
-  const resetEditorContent = () => {
-    const currentMentions = getEditorCurrentMentions();
-    let newEditorState = EditorState.createEmpty();
-    currentMentions.forEach((mention) => {
-      if (mention.trigger === "@") {
-        const correspondingMention = mentionItems["@"].find(
-          (item) => item.name === mention.content
-        );
-        if (correspondingMention !== undefined) {
-          if (mention.trigger === "@") {
-            newEditorState = insertMention(
-              correspondingMention,
-              mention.trigger,
-              newEditorState
-            );
-          }
-        }
-      }
-    });
-    setEditorState(newEditorState);
-  };
-
-  const getEditorCurrentMentions = (): TriggerMap[] => {
-    const contentState = editorState.getCurrentContent();
-    const plainText = contentState.getPlainText();
-    const mentionTriggers = Object.keys(mentionItems);
-
-    const mentionTexts: TriggerMap[] = [];
-
-    mentionTriggers.forEach((trigger) => {
-      if (trigger === "@") {
-        mentionItems["@"].forEach((item) => {
-          const mentionText = `${trigger}${item.name}`;
-          if (plainText.includes(mentionText)) {
-            mentionTexts.push({
-              trigger: trigger as MentionTriggerType,
-              content: item.name,
-            });
-          }
-        });
-      }
-    });
-    return mentionTexts;
-  };
-
-  const keyBindingFn = (e: React.KeyboardEvent<HTMLDivElement>) => {
-    if (e.key === "Enter" && !e.shiftKey) {
-      onSubmit();
-      return "submit";
-    }
-
-    return getDefaultKeyBinding(e);
-  };
-
-  const getNonMentionTexts = (editorCurrentState: EditorState): string => {
-    const contentState = editorCurrentState.getCurrentContent();
-    let plainText = contentState.getPlainText();
-    Object.keys(mentionItems).forEach((trigger) => {
-      if (trigger === "@") {
-        mentionItems[trigger].forEach((item) => {
-          const regex = new RegExp(`${trigger}${item.name}`, "g");
-          plainText = plainText.replace(regex, "");
-        });
-      }
-    });
-    return plainText;
-  };
-
-  const handleEditorChange = (newEditorState: EditorState) => {
-    setEditorState(newEditorState);
-    const currentMessage = getNonMentionTexts(newEditorState);
-    setMessage(currentMessage);
-  };
-
-  useEffect(() => {
-    if (message === "") {
-      resetEditorContent();
-    }
-  }, [message]);
 
   return (
     <div
@@ -151,12 +62,10 @@ export const MentionInput = ({
         onSearchChange={onSearchChange}
         popoverContainer={({ children }) => {
           return (
-            <Popover>
-              <div className="z-50 bg-white dark:bg-black border border-black/10 dark:border-white/25 rounded-md shadow-md overflow-y-auto min-w-32">
-                {children}
-                <AddNewBrainButton />
-              </div>
-            </Popover>
+            <div className="z-50 bg-white dark:bg-black border border-black/10 dark:border-white/25 rounded-md shadow-md overflow-y-auto min-w-32">
+              {children}
+              <AddNewBrainButton />
+            </div>
           );
         }}
         onAddMention={onAddMention}
