+++ frontend/app/chat/[chatId]/components/ActionsBar/components/ChatInput/components/ChatBar/components/MentionInput/hooks/helpers/MentionPlugin.tsx
@@ -24,6 +24,7 @@ export const useMentionPlugin = (props: MentionPluginProps) => {
           }}
         />
       ),
+      mentionPrefix: "@",
       popperOptions: {
         placement: "top-end",
       },
