+++ frontend/app/chat/[chatId]/components/ActionsBar/components/ChatInput/components/ChatBar/components/MentionInput/hooks/helpers/MentionState.ts
@@ -1,10 +1,10 @@
-/* eslint-disable max-lines */
 import { EditorState } from "draft-js";
 import { useEffect, useState } from "react";
 
 import { useBrainContext } from "@/lib/context/BrainProvider/hooks/useBrainContext";
 
-import { MentionInputMentionsType } from "../../../../types";
+import { MentionTriggerType } from "@/app/chat/[chatId]/components/ActionsBar/types";
+import { MentionInputMentionsType, TriggerMap } from "../../../../types";
 import { mapMinimalBrainToMentionData } from "../../utils/mapMinimalBrainToMentionData";
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
@@ -32,6 +32,45 @@ export const useMentionState = () => {
     legacySetEditorState(stateWithContentAndSelection);
   };
 
+  const getEditorCurrentMentions = (): TriggerMap[] => {
+    const contentState = editorState.getCurrentContent();
+    const plainText = contentState.getPlainText();
+    const mentionTriggers = Object.keys(mentionItems);
+
+    const mentionTexts: TriggerMap[] = [];
+
+    mentionTriggers.forEach((trigger) => {
+      if (trigger === "@") {
+        mentionItems["@"].forEach((item) => {
+          const mentionText = `${trigger}${item.name}`;
+          if (plainText.includes(mentionText)) {
+            mentionTexts.push({
+              trigger: trigger as MentionTriggerType,
+              content: item.name,
+            });
+          }
+        });
+      }
+    });
+    return mentionTexts;
+  };
+
+  const getEditorTextWithoutMentions = (
+    editorCurrentState: EditorState
+  ): string => {
+    const contentState = editorCurrentState.getCurrentContent();
+    let plainText = contentState.getPlainText();
+    Object.keys(mentionItems).forEach((trigger) => {
+      if (trigger === "@") {
+        mentionItems[trigger].forEach((item) => {
+          const regex = new RegExp(`${trigger}${item.name}`, "g");
+          plainText = plainText.replace(regex, "");
+        });
+      }
+    });
+    return plainText;
+  };
+
   useEffect(() => {
     setMentionItems({
       ...mentionItems,
@@ -44,11 +83,6 @@ export const useMentionState = () => {
     });
   }, [allBrains]);
 
-  useEffect(() => {
-    console.log("changed");
-    console.log({ editorState });
-  }, [editorState]);
-
   return {
     editorState,
     setEditorState,
@@ -56,5 +90,7 @@ export const useMentionState = () => {
     setSuggestions,
     setMentionItems,
     suggestions,
+    getEditorCurrentMentions,
+    getEditorTextWithoutMentions,
   };
 };
