+++ frontend/app/chat/[chatId]/components/ActionsBar/components/ChatInput/components/ChatBar/types.ts
@@ -1,5 +1,10 @@
 import { MinimalBrainForUser } from "@/lib/context/BrainProvider/types";
+import { MentionTriggerType } from "../../../../types";
 
 export type MentionInputMentionsType = {
   "@": MinimalBrainForUser[];
+};
+export type TriggerMap = {
+  trigger: MentionTriggerType;
+  content: string;
 };
