+++ frontend/lib/components/BrainUsers/hooks/useBrainUsers.ts
@@ -1,6 +1,5 @@
-/* eslint-disable max-lines */
 import { useEffect, useState } from "react";
-import { useTranslation } from 'react-i18next'
+import { useTranslation } from "react-i18next";
 
 import { Subscription } from "@/lib/api/brain/brain";
 import { useBrainApi } from "@/lib/api/brain/useBrainApi";
@@ -15,7 +14,7 @@ export const useBrainUsers = (brainId: string) => {
   const { publish } = useToast();
   const { getBrainUsers } = useBrainApi();
   const { session } = useSupabase();
-  const { t } = useTranslation(['brain']);
+  const { t } = useTranslation(["brain"]);
 
   const fetchBrainUsers = async () => {
     // Optimistic update
@@ -26,7 +25,7 @@ export const useBrainUsers = (brainId: string) => {
     } catch {
       publish({
         variant: "danger",
-        text: t('errorFetchingBrainUsers',{ns:'brain'})
+        text: t("errorFetchingBrainUsers", { ns: "brain" }),
       });
     } finally {
       setFetchingBrainUsers(false);
